refactored linter engine
